rabbit_celery_hw/requirements.txt
celery
fastapi
uvicorn
Pillow
python-multipart
redis